Update karate version to 1.2.0
Latest and greatest, and an actual release version (0.9.5 was still in development).
Does need some changes, and most noticable is the need to construct a ScenarioRuntime which turned out to require quite some instantiation as nullpointers would occur without it.
Karate is more meant for testing purposes, so the way we use it needs some extra work... but it works (at least for me locally).
Does require the timeout to be specified as an integer instead of a long (warning for at least AERIUS-II).